refactor: move Cookies to be an util instead of service
